+++ custom-endpoints/twitch.js
@@ -1,18 +1,19 @@
+import envHandler from '../environment_handler';
 // The Twitch API url to hit for EFT data
 const url = "https://api.twitch.tv/helix/streams?game_id=491931&first=100"
-// Twitch API headers for authentication
-const init = {
-    headers: {
-        'Authorization': `Bearer ${TWITCH_TOKEN}`,
-        'Client-ID': TWITCH_CLIENT_ID,
-    },
-};
+
 // Response content-type headers
 const contentType = 'application/json;charset=UTF-8';
 
 // Make a request to the Twitch API to get data for EFT streams
 async function getTwitchData() {
-    let response = await fetch(url, init);
+    let response = await fetch(url, {
+        // Twitch API headers for authentication
+        headers: {
+            'Authorization': `Bearer ${envHandler.getEnv().TWITCH_TOKEN}`,
+            'Client-ID': envHandler.getEnv().TWITCH_CLIENT_ID,
+        },
+    });
     let responseJson = JSON.stringify(await response.json(), null, 2);
     return responseJson;
 }
